Update and rename gf-completion.zsh to scan_metasploit.sh
To use this tool, save it to a file (e.g. scan_metasploit.sh) and make it executable with chmod +x scan_metasploit.sh. Then you can run it with ./scan_metasploit.sh



To use this tool, save it to a file (e.g. scan_metasploit.sh) and make it executable with chmod +x scan_metasploit.sh. Then you can run it with ./scan_metasploit.sh