Create SHELTP.sh
Sure, here is a simple script that can check the HTTP status code for a list of URLs or subdomains. The script reads each URL or subdomain from a file, one per line, and uses the curl command to send an HTTP request to each URL. The status code of the response is then printed to the console.



You can specify the path to the file containing the list of URLs or subdomains by replacing "url-list.txt" with the desired file path.